feat: add environ format
Adds a `direnv export environ` command that outputs the diff in
`<key>(=<value)?\0` format. It's a simple format that is quite close to
/proc/<pid>/environ on Linux. Each line is NUL-terminated. If the `=` is
not present on a line, consider that the <key> should be unset.

The idea was to find a format that could eventually replace the evals in
the hooks in the various shells.